requirements.txt
streamlit
faiss-cpu
numpy
torch
transformers
sentence-transformers
PyPDF2
gTTS
